Separate snapshot management from docker storage pool
storage.snapshot.create/backup/remove would be handled separately
